pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "wagtail-oidc-provider"
version = "0.0.0"
description = "Bridging django-oidc-provider to wagtail admin CMS"
authors = [
    {name = "WebU dev team", email = "contact@webu.coop"},
]
readme = "README.md"
requires-python = ">= 3.6"
dependencies = [
    "django>=2.2",
    "wagtail",
    "django-oidc-provider",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Framework :: Django",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
]
keywords=["django", "wagtail", "oidc", "sso"]

[project.urls]
Homepage = "https://github.com/webu/wagtail-oidc-provider"
Issues = "https://github.com/webu/wagtail-oidc-provider/issues"

[tool.ruff]
    include = ["pyproject.toml", "wagtail_oidc_provider/**/*.py"]
    line-length = 99
    extend-exclude = [
        "__pycache__",
        "migrations"
    ]

[tool.ruff.lint]
    select = [
        "F401", # unused imports
        "I",
    ] # https://docs.astral.sh/ruff/rules

[tool.ruff.lint.isort]
    length-sort = false
    force-single-line = true
    force-sort-within-sections = true
    section-order = ["future", "standard-library", "pylib", "django", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.isort.sections]
"django" = ["django"]

